ordine carrello
ordine carrello

+++ Spring/IdsFilieraAgricola/src/main/java/it/cs/unicam/ids/filiera/demo/entity/Distributore.java
@@ -3,6 +3,8 @@ package it.cs.unicam.ids.filiera.demo.entity;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 
+import java.math.BigDecimal;
+
 @Entity
 @DiscriminatorValue("DISTRIBUTORE")
 public class Distributore extends Venditore {
@@ -15,7 +17,7 @@ public class Distributore extends Venditore {
 	}
 
 	@Override
-	public Prodotto creaProdotto(String nomeProdotto, String categoria, double prezzo) {
+	public Prodotto creaProdotto(String nomeProdotto, String categoria, BigDecimal prezzo) {
 		return null;
 	}
