+++ Spring/IdsFilieraAgricola/src/main/java/it/cs/unicam/ids/filiera/demo/entity/Trasformatore.java
@@ -4,6 +4,8 @@ package it.cs.unicam.ids.filiera.demo.entity;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 
+import java.math.BigDecimal;
+
 @Entity
 @DiscriminatorValue("TRASFORMATORE")
 public class Trasformatore extends Venditore {
@@ -17,7 +19,7 @@ public class Trasformatore extends Venditore {
 	}
 
 	@Override
-	public Prodotto creaProdotto(String nomeProdotto, String categoria, double prezzo) {
+	public Prodotto creaProdotto(String nomeProdotto, String categoria, BigDecimal prezzo) {
 		return null;
 	}
 }
